+++ shared/src/random.c
@@ -27,14 +27,14 @@ unsigned int rand_i()
 	return gsl_rng_get(rng);
 }
 
-unsigned float rand_f()
+float rand_f()
 {
-	return (unsigned float) rand_d();
+	return (float) rand_d();
 }
 
-unsigned double rand_d()
+double rand_d()
 {
-	return gsl_rng_uniform(rng) + (unsigned double) rand_i();
+	return gsl_rng_uniform(rng) + (double) rand_i();
 }
 
 unsigned int rand_i_b(unsigned int max)
@@ -44,11 +44,11 @@ unsigned int rand_i_b(unsigned int max)
 
 // todo recheck if needed for validity
 
-unsigned double rand_d_b(unsigned double max)
+double rand_d_b(double max)
 {
 	unsigned int temp_i;
-	unsigned double temp_d;
-	unsigned double gen = 0;
+	double temp_d;
+	double gen = 0;
 	do {
 		temp_i = rand_i_b((unsigned int) max);
 		temp_d = rand_d();
@@ -56,7 +56,7 @@ unsigned double rand_d_b(unsigned double max)
 	return gen;
 }
 
-unsigned float rand_f_b(unsigned float max)
+float rand_f_b(float max)
 {
-	return (unsigned float) rand_d_b((unsigned double) max);
+	return (float) rand_d_b((double) max);
 }
